カテゴリー投稿・更新・削除のAPI制限

+++ src/app/admin/categories/[id]/page.tsx
@@ -6,17 +6,26 @@ import { useForm, SubmitHandler } from "react-hook-form";
 import { RequestCategoryBody } from "../../../_type/RequestCategoryBody";
 import { PostProps } from "@/app/_type/PostProps";
 import CategoryForm from "@/app/_components/CategoryForm";
+import useSupabaseSession from "@/app/_hooks/useSupabaseSession";
 
 const EditCategories: React.FC<PostProps> = ({params}) => {
   const { id } = params
   const router = useRouter();
   const [category, setCategory] = useState<RequestCategoryBody>();
   const { setValue } = useForm<RequestCategoryBody>();
+  const { token } = useSupabaseSession();
   
   useEffect(() => {
+    if (!token) return
+
     const findCategory = async () => {
       try{
-        const response = await fetch(`/api/admin/categories/${id}`)
+        const response = await fetch(`/api/admin/categories/${id}`, {
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: token,
+          },
+        })
         const data = await response.json();
         setCategory(data.category)
 
@@ -29,8 +38,14 @@ const EditCategories: React.FC<PostProps> = ({params}) => {
   }, []);
   
   const onsubmit: SubmitHandler<RequestCategoryBody> = async(data) => {
+    if (!token) return 
+
     try {
       const response = await fetch(`/api/admin/categories/${id}`, {
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: token,
+        },
         method: "PUT",
         body: JSON.stringify(data), 
       })
@@ -45,8 +60,14 @@ const EditCategories: React.FC<PostProps> = ({params}) => {
   }
 
   const handleDelete = async() => {
+    if (!token) return
+    
     try {
       const response = await fetch(`/api/admin/categories/${id}`, {
+        headers: {
+          "Content-type": "application/json",
+          Authorization: token,
+        },
         method: "DELETE",
         body: JSON.stringify(id)
       })
